factor OrderByTrait into OrderByInterface
