+++ src/Mysql/Delete.php
@@ -20,11 +20,10 @@ use Aura\Sql_Query\Traits;
  * @package Aura.Sql_Query
  *
  */
-class Delete extends Common\Delete
+class Delete extends Common\Delete implements Common\OrderByInterface
 {
     use Traits\LimitTrait;
-    use Traits\OrderByTrait;
-    
+
     /**
      * 
      * Converts this query object to a string.
@@ -83,5 +82,19 @@ class Delete extends Common\Delete
     {
         $this->setFlag('QUICK', $enable);
         return $this;
+    }
+
+    /**
+     *
+     * Adds a column order to the query.
+     *
+     * @param array $spec The columns and direction to order by.
+     *
+     * @return $this
+     *
+     */
+    public function orderBy(array $spec)
+    {
+        return $this->addOrderBy($spec);
     }
 }
