+++ src/Mysql/Update.php
@@ -20,11 +20,10 @@ use Aura\Sql_Query\Traits;
  * @package Aura.Sql_Query
  *
  */
-class Update extends Common\Update
+class Update extends Common\Update implements Common\OrderByInterface
 {
     use Traits\LimitTrait;
-    use Traits\OrderByTrait;
-    
+
     /**
      * 
      * Converts this query object to a string.
@@ -68,5 +67,19 @@ class Update extends Common\Update
     {
         $this->setFlag('IGNORE', $enable);
         return $this;
+    }
+
+    /**
+     *
+     * Adds a column order to the query.
+     *
+     * @param array $spec The columns and direction to order by.
+     *
+     * @return $this
+     *
+     */
+    public function orderBy(array $spec)
+    {
+        return $this->addOrderBy($spec);
     }
 }
