--- src/Traits/OrderByTrait.php
@@ -1,61 +1,0 @@
-<?php
-/**
- * 
- * This file is part of Aura for PHP.
- * 
- * @package Aura.Sql_Query
- * 
- * @license http://opensource.org/licenses/bsd-license.php BSD
- * 
- */
-namespace Aura\Sql_Query\Traits;
-
-/**
- * 
- * A trait for adding ORDER BY.
- * 
- * @package Aura.Sql_Query
- * 
- */
-trait OrderByTrait
-{
-    /**
-     *
-     * ORDER BY these columns.
-     *
-     * @var array
-     *
-     */
-    protected $order_by = [];
-
-    /**
-     *
-     * Adds a column order to the query.
-     *
-     * @param array $spec The columns and direction to order by.
-     *
-     * @return $this
-     *
-     */
-    public function orderBy(array $spec)
-    {
-        foreach ($spec as $col) {
-            $this->order_by[] = $this->quoteNamesIn($col);
-        }
-        return $this;
-    }
-
-    /**
-     * 
-     * Appends the `ORDER BY ...` clause to the statement.
-     * 
-     * @return null
-     * 
-     */
-    protected function buildOrderBy()
-    {
-        if ($this->order_by) {
-            $this->stm .= PHP_EOL . 'ORDER BY' . $this->indentCsv($this->order_by);
-        }
-    }
-}
